Inventory Management System/pyproject.toml
[project]
name = "inventory-management-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=45.0.5",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
]
